pyproject.toml
[tool.poetry]
name = "flightapi"
version = "0.1.0"
description = ""
authors = ["Andrew Cusick <andrewcusick13@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.91.0"
uvicorn = {extras = ["standard"], version = "^0.20.0"}
sqlalchemy-utils = "^0.40.0"
sqlmodel = "^0.0.8"
psycopg2-binary = "^2.9.5"


[tool.poetry.group.dev.dependencies]
pytest = "^7.2.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
